fix: reject non-finite purchase values and widen event_time to Int64
Addresses Gemini Code Assist review feedback on PR #25: value <= 0
doesn't reject NaN/Infinity (comparisons with NaN are always false,
so Kotlin's <= 0 check let NaN through as "valid"), and Int can
overflow event_time on 32-bit architectures.

+++ ios/Sources/PoltioSDK/PoltioAPIClient.swift
@@ -243,7 +243,7 @@ final class PoltioAPIClient {
             payload["currency"] = currency
         }
         if let eventTime {
-            payload["event_time"] = Int(eventTime.timeIntervalSince1970)
+            payload["event_time"] = Int64(eventTime.timeIntervalSince1970)
         }
         if !items.isEmpty {
             payload["contents"] = items.map { item -> [String: Any] in
